Reorganize code into packages
- cmd/thriftlint is just the executable
- lint/ defines the framework for writing linter rules
- lint/rule/ defines default rules for ThriftRW
