+++ cmd/thriftlint/main.go
@@ -5,41 +5,28 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
-	"sort"
 	"strings"
 
 	"go.uber.org/thriftrw/ast"
 	"go.uber.org/thriftrw/idl"
+	"go.uber.org/thriftrw/lint"
+	"go.uber.org/thriftrw/lint/rule"
 )
 
 type catcher struct {
-	lints []Lint
+	lints []lint.Lint
 }
 
-func (c *catcher) PostLint(l Lint) {
+func (c *catcher) PostLint(l lint.Lint) {
 	c.lints = append(c.lints, l)
 }
 
-type byLine []Lint
-
-func (ll byLine) Len() int {
-	return len(ll)
-}
-
-func (ll byLine) Less(i, j int) bool {
-	return ll[i].Line < ll[j].Line
-}
-
-func (ll byLine) Swap(i, j int) {
-	ll[i], ll[j] = ll[j], ll[i]
-}
-
-var _simpleRules = []*SimpleRule{
-	javaNamespace,
-	errorFieldRule,
-	unannotedI64Rule,
-	unspecifiedRequiredness,
-	tooManyArguments,
+var _rules = []lint.Rule{
+	rule.JavaNamespace,
+	rule.ErrorField,
+	rule.UnannotatedI64,
+	rule.UnspecifiedRequiredness,
+	rule.TooManyArguments,
 }
 
 func main() {
@@ -53,22 +40,17 @@ func main() {
 		log.Fatal(err)
 	}
 
-	var rules []*Rule
-	for _, r := range _simpleRules {
-		rules = append(rules, r.mustBuild())
-	}
-
 	var (
 		c        catcher
 		visitors []ast.Visitor
 	)
-	for _, rule := range rules {
+	for _, rule := range _rules {
 		visitors = append(visitors, rule.Inspector(&c))
 	}
 
 	ast.Walk(ast.MultiVisitor(visitors...), prog)
 
-	sort.Sort(byLine(c.lints))
+	lint.SortByLine(c.lints)
 	for _, lint := range c.lints {
 		prefix := fmt.Sprintf("%d: ", lint.Line)
 		indent := len(prefix)
