+++ lint/rule/doc.go
@@ -1,0 +1,3 @@
+// Package rule implements the different linter rules the ThriftRW linter knows
+// about.
+package rule
